pyproject.toml
[project]
name = "rag-tester"
version = "0.1.0"
description = "A Python project for testing and evaluating Retrieval-Augmented Generation (RAG) systems"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "Sebastien MORAND", email = "sebastien.morand@ibm.com" }
]

dependencies = [
    "typer>=0.15.0",
    "rich>=13.9.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "aiofiles>=23.0.0",
    "nest-asyncio>=1.5.0",
    "psycopg[binary]>=3.1.0",
    "pymilvus>=2.6.12",
    "aiosqlite>=0.22.1",
    "elasticsearch>=9.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.14.0",
    "ruff>=0.9.0",
    "bandit>=1.8.0",
    "pre-commit>=4.0.0",
    "types-aiofiles>=23.0.0",
    "types-PyYAML>=6.0.0",
    "pytest-mock>=3.15.1",
]

[project.scripts]
rag-tester = "rag_tester:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_tester"]

# Ruff (Linting + Formatting)
[tool.ruff]
target-version = "py313"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented code)
    "PL",     # Pylint
    "RUF",    # Ruff-specific rules
    "G",      # flake8-logging-format (e.g., G004 forbids f-strings in logger calls)
]
ignore = [
    "E501",    # Line too long (handled by formatter)
    "PLR0913", # Too many arguments
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "PLR2004", "PTH123", "PLW1510", "PLC0415", "W293", "ERA001", "E741"]
# Typer commands are inherently long (option declarations + orchestration). PLR* and
# B008 (Typer's typing.Annotated/typer.Option pattern) are pragmatic exemptions on the
# CLI surface only. PLC0415 stays only here because backend providers and nest_asyncio
# are imported lazily to keep the CLI startup fast and to allow the optional backends
# to be missing without crashing the whole CLI.
"src/rag_tester/commands/*.py" = ["PLR0911", "PLR0912", "PLR0915", "PLR2004", "RUF001", "B008", "PLC0415"]
"src/rag_tester/core/*.py" = ["PLR0912", "PLR0915", "PLR2004"]
# sentence-transformers is intentionally lazy-imported in local.py to avoid loading
# torch on every CLI startup.
"src/rag_tester/providers/embeddings/local.py" = ["PLC0415"]
"src/rag_tester/providers/*.py" = ["PLR2004"]
"src/rag_tester/utils/*.py" = ["PLR2004"]

[tool.ruff.lint.isort]
known-first-party = ["rag_tester"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# mypy (Type Checking)
[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = "opentelemetry.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "nest_asyncio"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "psycopg.*"
ignore_missing_imports = true

# pytest
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short", "--strict-markers"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "e2e: End-to-end tests",
    "critical: Critical path tests",
    "high: High priority tests",
    "medium: Medium priority tests",
    "low: Low priority tests",
]

# Coverage
[tool.coverage.run]
source = ["src"]
branch = true
omit = ["tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 80
show_missing = true
